fix(mux): preserve full reaper ancestry protection
Protect PID 1 and every ancestor after sharing the sweep helper, and keep sidebar unit tests in one documented sibling tree.
